requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
openai==1.3.8
python-socketio==5.10.0
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
